// apps/api/src/modules/openai/dto/completions-dtos/CompletionTokensDetailsDto.ts
import { ApiProperty } from '@nestjs/swagger';
import { IsNumber, IsOptional } from 'class-validator';

export class CompletionTokensDetailsDto {
  /**
   * When using Predicted Outputs, the number of tokens in the prediction that
   * appeared in the completion.
   */
  @ApiProperty({
    required: false,
    description: `When using Predicted Outputs, the number of tokens in the prediction that
  appeared in the completion.`,
    type: 'number',
  })
  @IsOptional()
  @IsNumber()
  accepted_prediction_tokens?: number;

  /** Audio input tokens generated by the model. */
  @ApiProperty({
    required: false,
    description: `Audio input tokens generated by the model.`,
    type: 'number',
  })
  @IsOptional()
  @IsNumber()
  audio_tokens?: number;

  /** Tokens generated by the model for reasoning. */
  @ApiProperty({
    required: false,
    description: `Tokens generated by the model for reasoning.`,
    type: 'number',
  })
  @IsOptional()
  @IsNumber()
  reasoning_tokens?: number;

  /**
   * When using Predicted Outputs, the number of tokens in the prediction that did
   * not appear in the completion. However, like reasoning tokens, these tokens are
   * still counted in the total completion tokens for purposes of billing, output,
   * and context window limits.
   */
  @ApiProperty({
    required: false,
    description: `When using Predicted Outputs, the number of tokens in the prediction that did
  not appear in the completion. However, like reasoning tokens, these tokens are
  still counted in the total completion tokens for purposes of billing, output,
  and context window limits.`,
    type: 'number',
  })
  @IsOptional()
  @IsNumber()
  rejected_prediction_tokens?: number;
}